requirements.txt
oracledb==2.0.1
python-dotenv==1.0.0